Fixing Windows 10 DCD compilation
Removing the explicit --arch=x86_mscoff option for win32 platform build of DCD allows the compilation to complete successfully on my Windows 10. As the default destination of the dub build is not the bin subfolder, had to manually adapt the generated user settings. Will propose a change to DCD dub.json, keeping dub build in line with make and build.bat.