remove function parameters grammar
It was not matched with it only working if { was on the same line as )
Because we can't match newlines to see if there is a { after ), just
remove this rule. Does not affect a lot of highlighting, as parameters
are already matched as regular variable definitions, which highlight
fine.