make dubCompiler setting machine-specific
For remote development this means the compiler must be changed on the
remote machine and won't be inherited from user settings.

However workspace / folder settings are unaffected and will still work.

This change is because dubCompiler is very platform specific and can
include full local paths.

Sync extensions might also respect this.